pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "My original utility functions"
authors = [
    { name = "n4okins", email = "63445867+n4okins@users.noreply.github.com" }
]
dependencies = [
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "pytest>=8.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/utils"]
